refactor(cli): split runtime support commands
